adicionei novas fontes e parte de comentario diminuiu

+++ Quadro/PosCameraViewController.m
@@ -89,8 +89,8 @@
         int cont = (int) assuntoE.listaFotosComAnotacao.count+1;
         [assuntoE.listaFotosComAnotacao addObjectsFromArray:self.listaDeFotosComAnotacao];
         for (FotoComAnotacao *foto in self.listaDeFotosComAnotacao) {
-            [foto nomeDaFotoAssunto:assuntoE posicaoFoto:cont++ idMateria:self.posicaoMateria];
-            [self gravaFoto:foto doAssunto:assuntoE comIdMateria:self.posicaoMateria];
+            [foto nomeDaFotoAssunto:assuntoE posicaoFoto:cont++ idMateria:(int)self.posicaoMateria];
+            [self gravaFoto:foto doAssunto:assuntoE comIdMateria:(int)self.posicaoMateria];
         }
     } else {
         Assunto *assunto = [[Assunto alloc] initAssuntoPorData:[NSDate date] comNomeAssunto:self.txtMateria.text];
@@ -100,8 +100,8 @@
         // VAI TER QUE SALVAR AQUI
         int cont = 1;
         for (FotoComAnotacao *foto in self.listaDeFotosComAnotacao) {
-            [foto nomeDaFotoAssunto:assunto posicaoFoto:cont++ idMateria:self.posicaoMateria];
-            [self gravaFoto:foto doAssunto:assunto comIdMateria:self.posicaoMateria];
+            [foto nomeDaFotoAssunto:assunto posicaoFoto:cont++ idMateria:(int)self.posicaoMateria];
+            [self gravaFoto:foto doAssunto:assunto comIdMateria:(int)self.posicaoMateria];
         }
     }
     [self dismissViewControllerAnimated:YES completion:nil];
